2026-01-demo-v0.5/fastapi-inventory/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
arcosparse==0.4.2
boto3==1.40.66
botocore==1.40.66
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.2
colorama==0.4.6
copernicusmarine==2.2.4
crc32c==2.8
dask==2025.10.0
dnspython==2.8.0
donfig==0.8.1.post1
fastapi==0.115.2
fsspec==2025.10.0
h11==0.16.0
h5netcdf==1.7.3
h5py==3.15.1
httptools==0.7.1
idna==3.11
Jinja2==3.1.4
jmespath==1.0.1
locket==1.0.0
MarkupSafe==3.0.3
motor==3.6.0
numcodecs==0.16.3
numpy==2.3.4
packaging==25.0
pandas==2.3.3
partd==1.4.2
pyarrow==22.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.2
pystac==1.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.14.0
semver==3.0.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.40.0
toolz==1.1.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.30.6
watchfiles==1.1.1
websockets==15.0.1
xarray==2025.10.1
zarr==3.1.3
cfgrib==0.9.13.0
